Restructure to avoid circular dependencies.
