+++ src/Instruction.h
@@ -8,63 +8,10 @@
 #include <typeinfo>
 #include <stdio.h>
 
+#include "IrEnums.h"
+#include "typedefs.h"
+#include "BasicBlock.h"
 #include "StorageLocation.h"
-
-enum ValType {VAR, ARRAY, LITERAL, EMPTY};
-enum DataType {FLOAT, INT, VOID, UNKNOWN};
-
-struct ProgramValue {
-    ValType vtype;
-    DataType dtype;
-    std::string value;
-    int size;
-};
-
-enum InstOpType 
-{
-    ASSIGN,
-    ADD,
-    SUB,
-    MULT,
-    DIV,
-    AND,
-    OR,
-    GOTO,
-    BREQ,
-    BRNEQ,
-    BRLT,
-    BRGT,
-    BRLEQ,
-    BRGEQ,
-    RETURN_VOID,
-    RETURN_NONVOID,
-    CALL,
-    CALLR,
-    ARRAY_STORE,
-    ARRAY_LOAD,
-    ARRAY_ASSIGN
-};
-
-namespace IR {
-    // Forward declaration
-    class BasicBlock;
-
-    enum InstType
-    {
-        BINARY,
-        ASSIGN,
-        BRANCH,
-        RETURN,
-        CALL,
-        ARRAY
-    };
-}; //end namespace IR
-
-// Forward declaration
-class Instruction;
-
-
-typedef std::vector<Instruction*> InstContainer;
 
 
 class Instruction {
@@ -91,6 +38,7 @@ class Instruction {
         void addPredecessor(Instruction* predecessor);
         void printSuccessors();
         void printPredecessors();
+        void printParent();
 
         InstContainer* getSuccessors();
         InstContainer* getPredecessors();
@@ -100,7 +48,7 @@ class Instruction {
         bool isLeader() { return leader; };
         void isBranchTarget(bool brTarget) { branchTarget = brTarget; }
         bool isBranchTarget() { return branchTarget; }
-        void setParent(IR::BasicBlock* parent); { parent_ = parent; }
+        void setParent(IR::BasicBlock* parent) { parent_ = parent; }
         IR::BasicBlock* getParent() { return parent_; }
 
         //pure virtual
